Fixes for openwrt compilation. Fix crash bug in getStateString. Close serial port when printer is disconnected. Fix off-by-one bug causing newlines after comments to disappear.

+++ src/server/main.cpp
@@ -78,7 +78,7 @@ int main(int argc, char** argv) {
 
 		ipc_free_device_list(devlist);
 	} else {
-		int lastSlash = serialDevice.rfind('/');
+		size_t lastSlash = serialDevice.rfind('/');
 		if (lastSlash != string::npos) serialDevice = serialDevice.substr(lastSlash + 1);
 	}
